refactor: extract pure matching helpers into matching.py
Move the pure, I/O-free matching functions (normalize, simple, isrc/duration/
name_match, artists_overlap, match, test_album_similarity, _names_match,
album_match, artist_match) out of the ~1000-line sync.py into a dedicated
matching.py leaf module, consolidating logic that was scattered across two
regions. sync.py re-exports the names so existing imports (and the test suite)
keep working unchanged. No behaviour change.
